Changed the coordinate type to f64

+++ crates/core/tests/test_pyxel.rs
@@ -1,13 +1,13 @@
 use pyxel::{Pyxel, PyxelCallback};
 
 pub struct App {
-    x: i32,
-    y: i32,
+    x: f64,
+    y: f64,
 }
 
 impl App {
     fn new(pyxel: &mut Pyxel) -> App {
-        let app = App { x: 0, y: 0 };
+        let app = App { x: 0.0, y: 0.0 };
 
         pyxel.mouse(true);
         pyxel.set_btnv(pyxel::MOUSE_POS_X, 0);
@@ -73,8 +73,8 @@ impl App {
 impl PyxelCallback for App {
     fn update(&mut self, pyxel: &mut Pyxel) {
         if pyxel.frame_count() < 60 {
-            self.x += (pyxel.frame_count() % 2) as i32;
-            self.y -= 1;
+            self.x += (pyxel.frame_count() % 2) as f64;
+            self.y -= 1.0;
         }
 
         if pyxel.btnp(pyxel::KEY_Q, None, None) {
@@ -84,11 +84,11 @@ impl PyxelCallback for App {
 
     fn draw(&mut self, pyxel: &mut Pyxel) {
         pyxel.cls(3);
-        pyxel.pset(self.x, 20, 7);
-        pyxel.rect(self.x + 10, 25, 15, 10, 8);
-        pyxel.rectb(self.x + 15, 45, 15, 10, pyxel::COLOR_WHITE);
+        pyxel.pset(self.x, 20.0, 7);
+        pyxel.rect(self.x + 10.0, 25.0, 15.0, 10.0, 8);
+        pyxel.rectb(self.x + 15.0, 45.0, 15.0, 10.0, pyxel::COLOR_WHITE);
 
-        pyxel.blt(0, 0, 0, 0, 0, 8, 8, None);
+        pyxel.blt(0.0, 0.0, 0, 0.0, 0.0, 8.0, 8.0, None);
     }
 }
